refactor: create CommentsFilterPopover
